+++ CHANGELOG.md
@@ -8,7 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Documentation
    - Added examples for JRO and GNSS data
 - Testing
-   - Added unit tests for JRO and GNSS methods
+   - Added unit tests for general, JRO, and GNSS methods
 
 ## [0.0.4] - 2021-06-11
 - Made changes to structure to comply with updates in pysat 3.0.0